basic maths operations added with gradient tracker included. -- fix error-- in this prototype the program doesn't optimized for speed.
